Add README.md + rename index to config + add middleware logic in its own file in the i18n folder
