+++ src/i18n/client.tsx
@@ -1,4 +1,4 @@
-import { LocalesI18nConfig } from "@/i18n";
+import { LocalesI18nConfig } from "@/i18n/config";
 import { createI18nClient } from "next-international/client";
 
 /**
